Remove duplicate tests and consolidate coverage
