--- test/browser/fetchAndCacheBlogData.status.additional.test.js
@@ -1,21 +1,0 @@
-import { describe, it, expect, jest } from '@jest/globals';
-import { fetchAndCacheBlogData } from '../../src/browser/data.js';
-
-describe('fetchAndCacheBlogData status transitions', () => {
-  it('sets status to loading then loaded on success', async () => {
-    const state = {
-      blog: null,
-      blogStatus: 'idle',
-      blogError: null,
-      blogFetchPromise: null,
-    };
-    const fetchFn = jest.fn(() =>
-      Promise.resolve({ ok: true, json: () => Promise.resolve({}) })
-    );
-    const loggers = { logInfo: jest.fn(), logError: jest.fn() };
-    const promise = fetchAndCacheBlogData(state, fetchFn, loggers);
-    expect(state.blogStatus).toBe('loading');
-    await promise;
-    expect(state.blogStatus).toBe('loaded');
-  });
-});
